+++ lib/Service/Places.php
@@ -43,12 +43,12 @@ class Places
         }
 
         // Detect database type
-        $platform = strtolower($this->connection->getDatabasePlatform()::class);
+        $platform = $this->getDbName();
 
         // Test MySQL-like support in databse
         if (str_contains($platform, 'mysql') || str_contains($platform, 'mariadb')) {
             try {
-                $res = $this->connection->executeQuery("SELECT ST_GeomFromText('POINT(1 1)')")->fetch();
+                $res = $this->connection->executeQuery("SELECT ST_GeomFromText('POINT(1 1)', 4326)")->fetch();
                 if (0 === \count($res)) {
                     throw new \Exception('Invalid result');
                 }
@@ -98,10 +98,12 @@ class Places
 
         // Construct WHERE clause depending on GIS type
         $where = null;
-        if (1 === $gisType) {
-            $where = "ST_Contains(geometry, ST_GeomFromText('POINT({$lon} {$lat})'))";
-        } elseif (2 === $gisType) {
-            $where = "POINT('{$lon},{$lat}') <@ geometry";
+        if (GIS_TYPE_MYSQL === $gisType) {
+            $where = "ST_Contains(geometry, ST_GeomFromText('POINT({$lat} {$lon})', 4326))";
+        } elseif (GIS_TYPE_POSTGRES === $gisType) {
+            // Postgres does not support using an index with POINT <@ POLYGON
+            // https://www.postgresql.org/docs/current/gist-builtin-opclasses.html
+            $where = "POLYGON('{$lat},{$lon}') <@ geometry";
         } else {
             return [];
         }
@@ -223,7 +225,7 @@ class Places
         $query = $this->connection->getQueryBuilder();
         $geomParam = (string) $query->createParameter('geometry');
         if (GIS_TYPE_MYSQL === $gis) {
-            $geomParam = "ST_GeomFromText({$geomParam})";
+            $geomParam = "ST_GeomFromText({$geomParam}, 4326)";
         } elseif (GIS_TYPE_POSTGRES === $gis) {
             $geomParam = "POLYGON({$geomParam}::text)";
         }
@@ -325,27 +327,35 @@ class Places
                     ++$idx;
                     $geometry = '';
 
+                    // Every polygon must have at least 3 points
                     if (\count($coords) < 3) {
                         echo "ERROR: Invalid polygon {$polyid}\n";
 
                         continue;
                     }
 
+                    // Check if coordinates are valid
+                    foreach ($coords as [$lon, $lat]) {
+                        if ($lon < -180 || $lon > 180 || $lat < -90 || $lat > 90) {
+                            echo "ERROR: Invalid coordinates for polygon {$polyid}\n";
+
+                            continue 2;
+                        }
+                    }
+
                     if (GIS_TYPE_MYSQL === $gis) {
                         $points = implode(',', array_map(static function (array $point) {
-                            $x = $point[0];
-                            $y = $point[1];
+                            [$lon, $lat] = $point;
 
-                            return "{$x} {$y}";
+                            return "{$lat} {$lon}";
                         }, $coords));
 
                         $geometry = "POLYGON(({$points}))";
                     } elseif (GIS_TYPE_POSTGRES === $gis) {
                         $geometry = implode(',', array_map(static function (array $point) {
-                            $x = $point[0];
-                            $y = $point[1];
+                            [$lon, $lat] = $point;
 
-                            return "({$x},{$y})";
+                            return "({$lat},{$lon})";
                         }, $coords));
                     }
 
@@ -382,7 +392,7 @@ class Places
         // Mark success
         echo "Planet database imported successfully!\n";
         flush();
-        $this->config->setSystemValue('memories.gis_type', $gis);
+        SystemConfig::set('memories.gis_type', $gis);
 
         // Delete data file
         @unlink($datafile);
@@ -427,14 +437,18 @@ class Places
             // Drop the table if it exists
             $this->connection->executeStatement('DROP TABLE IF EXISTS memories_planet_geometry');
 
+            // MySQL requires an SRID definition
+            // https://github.com/pulsejet/memories/issues/1067
+            $srid = str_contains($this->getDbName(), 'mysql') ? 'SRID 4326' : '';
+
             // Create table
-            $sql = 'CREATE TABLE memories_planet_geometry (
+            $sql = "CREATE TABLE memories_planet_geometry (
                 id varchar(32) NOT NULL PRIMARY KEY,
                 poly_id varchar(32) NOT NULL,
                 type_id int NOT NULL,
                 osm_id int NOT NULL,
-                geometry polygon NOT NULL
-            );';
+                geometry polygon NOT NULL {$srid}
+            );";
             $this->connection->executeQuery($sql);
 
             // Add indexes
@@ -444,10 +458,19 @@ class Places
             if (GIS_TYPE_MYSQL === $gis) {
                 $this->connection->executeQuery('CREATE SPATIAL INDEX planet_osm_polygon_geometry_idx ON memories_planet_geometry (geometry);');
             } elseif (GIS_TYPE_POSTGRES === $gis) {
-                $this->connection->executeQuery('CREATE INDEX planet_osm_polygon_geometry_idx ON memories_planet_geometry USING GIST (geometry);');
+                // https://www.postgresql.org/docs/current/gist-builtin-opclasses.html
+                $this->connection->executeQuery('CREATE INDEX planet_osm_polygon_geometry_idx ON memories_planet_geometry USING GIST (geometry poly_ops);');
             }
         } catch (\Exception $e) {
             throw new \Exception('Failed to create database tables: '.$e->getMessage());
         }
+    }
+
+    /**
+     * Get the database platform name.
+     */
+    private function getDbName(): string
+    {
+        return strtolower($this->connection->getDatabasePlatform()::class);
     }
 }