+++ src/components/modal/NodeShareModal.vue
@@ -17,11 +17,21 @@
 
     <div class="links">
       <ul>
-        <NcListItem v-for="share of shares" :name="share.label || t('memories', 'Share link')" :key="share.id"
-          :bold="false" :href="share.url" :compact="true" @click.prevent="shareOrCopy(share.url)">
+        <NcListItem
+          v-for="share of shares"
+          :title="share.label || t('memories', 'Share link')"
+          :name="share.label || t('memories', 'Share link')"
+          :key="share.id"
+          :bold="false"
+          :href="share.url"
+          :compact="true"
+          @click.prevent="shareOrCopy(share.url)"
+        >
           <template #icon>
             <LinkIcon class="avatar" :size="20" />
           </template>
+
+          <!-- NC_VUE: these slots have changed in v8 -->
           <template #subtitle>
             {{ getShareLabels(share) }}
           </template>