+++ src/typings/config.d.ts
@@ -26,6 +26,7 @@ declare module '@typings' {
     high_res_cond_default: HighResCond;
     livephoto_autoplay: boolean;
     sidebar_filepath: boolean;
+    metadata_in_slideshow: boolean;
 
     // folder settings
     folders_path: string;